+++ utils/storage/createUri.js
@@ -1,12 +1,3 @@
-/*
-
-	node-utils by SWR Audio Lab
-	frequently used node-packages and helpers
-
-	This module provides easy access to combine bucket + path to unique URIs
-
-*/
-
 module.exports = {
 	s3: (bucket, path) => {
 		// return data